Fork parallel-ssh 2.3.1 to Parallax SSH 1.0
To further the development of the pssh API over the command line tools
provided by pssh and due to the difficulty I have had in getting my API
patches merged upstream, I have decided to fork the project. The intent
of this project is somewhat different from that of PSSH - The goal is to
have a great API that is easy to use from other software, and to enable
the creation of tools like pssh and pnuke via python commands rather
than by providing command line tools directly.

The upstream for this project is https://github.com/krig/parallax-ssh
for now.